SF_Discord_Bot/requirements.txt
discord.py==2.0.0
requests>=2.25.0
python-dotenv
numpy